final_project_inf373/requirements.txt
asgiref==3.8.1
attrs==25.3.0
boto3==1.38.11
botocore==1.38.11
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
psycopg==3.2.7
psycopg-binary==3.2.7
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
rpds-py==0.24.0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
django-silk==5.3.2